+++ tests/display_address_test.rs
@@ -63,7 +63,7 @@ fn build_display_sim(builtin_name: &str, sub_inputs: &[(i32, PinBitCount)], sub_
 /// combinational propagation (and, for the last step of a phase, a clock edge) settle.
 fn drive(sim: &mut Simulator, values: &[(i32, u32)], steps: usize) {
 	let inputs: Vec<ExternalInput> =
-		values.iter().map(|&(id, state)| ExternalInput { address: PinAddress::new(id, id), state: PinState::from_raw(state) }).collect();
+		values.iter().map(|&(id, state)| ExternalInput { address: PinAddress::new(id, id), state: PinState::from_raw(state as u16) }).collect();
 	for _ in 0..steps {
 		sim.run_simulation_step(&inputs, &mut logic_sim::audio::SimAudio::new());
 	}
@@ -71,7 +71,7 @@ fn drive(sim: &mut Simulator, values: &[(i32, u32)], steps: usize) {
 
 fn read(sim: &Simulator, out_pin_id: i32) -> u32 {
 	let pin = sim.find_pin(sim.root(), PinAddress::new(out_pin_id, out_pin_id)).expect("wrapper output pin should resolve");
-	sim.pin(pin).state.raw()
+	sim.pin(pin).state.raw() as u32
 }
 
 // ---- DOT DISPLAY ----
